+++ test/unit/core/dom/test-dom.js
@@ -15,9 +15,8 @@
  */
 
 import * as dom from '#core/dom';
-import {createElementWithAttributes} from '#core/dom';
 import {loadPromise} from '../../../../src/event-helper';
-import {matches, realChildElements, realChildNodes} from '#core/dom/query';
+import {matches} from '#core/dom/query';
 import {setScopeSelectorSupportedForTesting} from '#core/dom/css-selectors';
 import {setShadowDomSupportedVersionForTesting} from '#core/dom/web-components';
 
@@ -503,39 +502,6 @@ describes.sandboxed('DOM', {}, (env) => {
       const focusSpy = env.sandbox.spy(element, 'focus');
       expect(() => dom.tryFocus(element)).to.not.throw();
       expect(focusSpy).to.have.been.called;
-    });
-  });
-
-  describe('realChildElements and realChildNodes', () => {
-    let element;
-    beforeEach(() => {
-      element = document.createElement('div');
-    });
-
-    it('realChildElements should return nothing', () => {
-      expect(realChildNodes(element).length).to.equal(0);
-      expect(realChildElements(element).length).to.equal(0);
-    });
-
-    it('realChildElements should return content-only nodes', () => {
-      const createWithAttr = (attr) =>
-        createElementWithAttributes(document, 'div', {[attr]: ''});
-
-      element.appendChild(document.createElement('i-amp-service'));
-      element.appendChild(createWithAttr('placeholder'));
-      element.appendChild(createWithAttr('fallback'));
-      element.appendChild(createWithAttr('overflow'));
-      element.appendChild(document.createTextNode('abc'));
-      element.appendChild(document.createElement('content'));
-
-      const nodes = realChildNodes(element);
-      expect(nodes.length).to.equal(2);
-      expect(nodes[0].textContent).to.equal('abc');
-      expect(nodes[1].tagName.toLowerCase()).to.equal('content');
-
-      const elements = realChildElements(element);
-      expect(elements.length).to.equal(1);
-      expect(elements[0].tagName.toLowerCase()).to.equal('content');
     });
   });
